refactor: hooks 폴더 내 eslint 오류 수정 및  커스텀 훅 파일 이동
-  hooks/lib/hooks 폴더로 커스텀 훅 파일 이동
