+++ hooks/src/__tests__/CardExpiryData.test.tsx
@@ -1,5 +1,6 @@
 import '@testing-library/jest-dom';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
+
 import CardExpiryDate from '../components/CardExpiryDate';
 import { getNowYearAndMonth } from '../lib/utils';
