+++ hooks/src/__tests__/CardIssuer.test.tsx
@@ -1,5 +1,5 @@
 import '@testing-library/jest-dom';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 
 import CardIssuer from '../components/CardIssuer';
