+++ hooks/src/lib/hooks/cardExpiryDate/useExpiryDate.ts
@@ -1,9 +1,11 @@
 import { useEffect, useState } from 'react';
-import { DATE_LENGTH, INVALID_INPUT_VALUE } from './constants';
-import { UseCardModuleReturn } from './types';
+
+import { DATE_LENGTH, INVALID_INPUT_VALUE } from '../../constants';
+import { UseCardModuleReturn } from '../../types';
+
+import useCardAvailability from './useCardAvailability';
 import useMonth from './useMonth';
 import useYear from './useYear';
-import useCardAvailability from './useCardAvailability';
 
 export interface ExpiryDate {
   month: string;
