+++ hooks/src/lib/hooks/cardNumbers/useCardNumbers.ts
@@ -1,36 +1,48 @@
 import { INVALID_INPUT_VALUE } from '../../constants';
 import { UseCardModuleReturn } from '../../types';
 import { sliceText } from '../../utils';
+
 import useCardNumbersValidation from './useCardNumbersValidation';
 import useCardNumbersValidationErrorMessage from './useCardNumbersValidationErrorMessage';
 
 export type CardNumbersType = (string | undefined)[];
-
+/**
+ * useCardNumbers props인 errorMessage 타입
+ */
+export interface UseCardNumbersErrorMessage {
+  empty: string;
+  number: string;
+  length: string;
+}
 export interface UseCardNumbersProps {
   fieldCount: number;
   cardNumberCounts: number[];
   cardNumbers: CardNumbersType;
-  errorMessages: {
-    empty: string;
-    number: string;
-    length: string;
-  };
+  errorMessages: UseCardNumbersErrorMessage;
   isNeedValidValue: boolean;
 }
 
 export type CardNumberError = 'empty' | 'number' | 'length' | null;
 
-export interface CardNumberValidation {
+/**
+ * 카드 번호 입력 필드 하나에 대한 유효성 검사 결과
+ */
+export interface CardNumberValidationResult {
   error: CardNumberError;
   isValid: boolean;
 }
-
-export type CardNumbersValidationResult = CardNumberValidation[];
-export type CardNumberValidationErrorMessage = (string | null)[];
+/**
+ * 모든 카드 번호 입력 필드에 대한 유효성 검사 결과
+ */
+export type CardNumbersValidationResults = CardNumberValidationResult[];
+/**
+ * useCardNumbers 가 반환하는 validationResultErrorMessage의 타입,모든 카드 번호 입력 필드의 오류 메세지가 배열로 담김
+ */
+export type CardNumberValidationResultErrorMessage = (string | null)[];
 
 export type UseCardNumbersReturn = UseCardModuleReturn<
-  CardNumberValidationErrorMessage,
-  CardNumbersValidationResult,
+  CardNumberValidationResultErrorMessage,
+  CardNumbersValidationResults,
   CardNumbersType
 >;
