+++ hooks/src/lib/hooks/useCVC.ts
@@ -1,45 +1,43 @@
 import { useEffect, useState } from 'react';
 
-import { validateFilledValue, validateNumber, validateLength, isValid } from './utils/validators';
-import { ErrorMessage, UseCardModuleReturn } from './types';
+import { INVALID_INPUT_VALUE } from '../constants';
+import { ErrorMessage, UseCardModuleReturn } from '../types';
+import { isValid, sliceText, validateFilledValue, validateLength, validateNumber } from '../utils';
 
-import { sliceText } from './utils/textFormatter';
-import { INVALID_INPUT_VALUE } from './constants';
-
-export interface PasswordValidationErrorMessages {
+export interface CardCVCValidationErrorMessages {
   empty: string;
   number: string;
   length: string;
 }
-export interface UsePasswordProps {
-  cardPassword: string;
-  errorMessages: PasswordValidationErrorMessages;
+export interface UseCardCVCProps {
+  cardCVC: string;
+  errorMessages: CardCVCValidationErrorMessages;
   validation: {
     length: number;
   };
   isNeedValidValue: boolean;
 }
-export interface UsePasswordResult {
+export interface UseCardCVCResult {
   isFilledValue: boolean;
   isValidNumber: boolean;
   isValidLength: boolean;
 }
 
-export type UsePasswordReturn = UseCardModuleReturn<ErrorMessage, UsePasswordResult, string>;
+export type UseCardCVCReturn = UseCardModuleReturn<ErrorMessage, UseCardCVCResult, string>;
 
-export default function usePassword(props: UsePasswordProps) {
-  const { cardPassword, errorMessages, validation, isNeedValidValue } = props;
+export default function useCVC(props: UseCardCVCProps): UseCardCVCReturn {
+  const { cardCVC, errorMessages, validation, isNeedValidValue } = props;
 
   type ErrorMessageKey = keyof typeof errorMessages;
   type Error = ErrorMessageKey[] | null;
   const [error, setError] = useState<Error>(null);
 
-  const validatePassword = () => {
+  const validateCVC = () => {
     const newError: ErrorMessageKey[] = [];
 
-    if (!validateFilledValue(cardPassword)) newError.push('empty');
-    if (!validateNumber(cardPassword)) newError.push('number');
-    if (!validateLength(cardPassword, validation.length)) newError.push('length');
+    if (!validateFilledValue(cardCVC)) newError.push('empty');
+    if (!validateNumber(cardCVC)) newError.push('number');
+    if (!validateLength(cardCVC, validation.length)) newError.push('length');
 
     setError(newError[0] ? newError : null);
   };
@@ -50,8 +48,8 @@ export default function usePassword(props: UsePasswordProps) {
    * @param error 유효성 검사 결과
    * @returns 변경된 글자
    */
-  const getFormattedValue = () => {
-    const slicedText = sliceText(cardPassword, validation.length);
+  const getFormattedValue = (value: string) => {
+    const slicedText = sliceText(value, validation.length);
     const isOnlyLengthError = error?.[0] === 'length';
 
     if (isNeedValidValue) {
@@ -62,8 +60,8 @@ export default function usePassword(props: UsePasswordProps) {
   };
 
   useEffect(() => {
-    validatePassword();
-  }, [cardPassword]);
+    validateCVC();
+  }, [cardCVC]);
 
   return {
     validationErrorMessage: error ? errorMessages[error[0]] : null,
@@ -72,6 +70,6 @@ export default function usePassword(props: UsePasswordProps) {
       isValidNumber: isValid<ErrorMessageKey>('number', error),
       isValidLength: isValid<ErrorMessageKey>('length', error),
     },
-    formattedValue: getFormattedValue(),
+    formattedValue: getFormattedValue(cardCVC),
   };
 }
