+++ hooks/src/lib/index.ts
@@ -1,7 +1,10 @@
-export { default as useCardHolder } from './useCardHolder';
-export { default as useCVC } from './useCVC';
-export { default as usePassword } from './usePassword';
-export { default as useCardIssuer } from './useCardIssuer';
-export { default as useCardType } from './useCardType';
-export { default as useExpiryDate } from './useExpiryDate';
+export { default as useCardHolder } from './hooks/useCardHolder';
+export { default as useCVC } from './hooks/useCVC';
+export { default as usePassword } from './hooks/usePassword';
+export { default as useCardIssuer } from './hooks/useCardIssuer';
+export { default as useCardType } from './hooks/useCardType';
+export { default as useExpiryDate } from './hooks/cardExpiryDate/useExpiryDate';
+export { default as useMonth } from './hooks/cardExpiryDate/useMonth';
+export { default as useYear } from './hooks/cardExpiryDate/useYear';
+export { default as useCardAvailability } from './hooks/cardExpiryDate/useCardAvailability';
 export { default as useCardNumbers } from './hooks/cardNumbers/useCardNumbers';
